# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinGuide / Финансовый капитал API",
    "version": "0.1.0",
    "description": "Backend/frontend contract for the financial planning product reconstructed from the Figma Make mockup. Base path: /api/v1. Current deployed real backend is exposed at /finguide-api/api/v1 on 66.42.121.18. Excel model analytics are documented in docs/model-analytics.md. Identity/auth is handled by Keycloak OIDC Authorization Code + PKCE; FinGuide API validates Bearer JWT and does not own password register/login endpoints."
  },
  "servers": [
    {
      "url": "http://66.42.121.18/finguide-api/api/v1",
      "description": "deployed real Spring Boot backend on 66.42.121.18"
    },
    {
      "url": "/api/v1",
      "description": "same-origin production API inside the backend service"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/plans/current": {
      "get": {
        "tags": [
          "Plan"
        ],
        "summary": "Get full normalized current plan state",
        "description": "Anonymous demo mode returns the seeded demo plan. Authenticated users get an idempotently created user-owned current plan cloned from the persisted demo seed on first access.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PlanState"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "put": {
        "tags": [
          "Plan"
        ],
        "summary": "Replace current plan state, used for migration/import",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanState"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PlanState"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/plans/{planId}/pension": {
      "parameters": [
        {
          "name": "planId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": [
          "Pension"
        ],
        "summary": "Get pension settings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PensionSettings"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "patch": {
        "tags": [
          "Pension"
        ],
        "summary": "Update pension settings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PensionSettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PensionSettings"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/plans/{planId}/dashboard": {
      "parameters": [
        {
          "name": "planId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Dashboard metrics matching mockup cards",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/DashboardMetrics"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/plans/{planId}/analytics/projection": {
      "parameters": [
        {
          "name": "planId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "years",
          "in": "query",
          "schema": {
            "type": "integer",
            "minimum": 1,
            "maximum": 60,
            "default": 30
          }
        }
      ],
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Long-term financial projection",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/YearlyProjectionPoint"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/scenarios": {
      "get": {
        "tags": [
          "Scenarios"
        ],
        "summary": "List user scenarios",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Scenario"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Scenarios"
        ],
        "summary": "Create scenario from blank plan, clone or adjustments",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Scenario"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Scenario"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/notifications/read": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Mark notifications as read",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "all": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/plans/{planId}/analytics/assumptions": {
      "parameters": [
        {
          "name": "planId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get projection assumptions imported from UI or Excel model",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ModelAssumptions"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "patch": {
        "tags": [
          "Analytics"
        ],
        "summary": "Update projection assumptions",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelAssumptions"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ModelAssumptions"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/plans/{planId}/analytics/balance/current": {
      "parameters": [
        {
          "name": "planId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Current-year workbook-style balance snapshot",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/BalanceSnapshot"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/plans/{planId}/analytics/cashflow": {
      "parameters": [
        {
          "name": "planId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "startYear",
          "in": "query",
          "schema": {
            "type": "integer"
          }
        },
        {
          "name": "endYear",
          "in": "query",
          "schema": {
            "type": "integer"
          }
        },
        {
          "name": "years",
          "in": "query",
          "schema": {
            "type": "integer",
            "minimum": 1,
            "maximum": 80
          }
        }
      ],
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Rich yearly cashflow projection based on the Excel model",
        "description": "Returns income, expense, goal-spending, annual savings and accumulated capital per year. API outflows are positive; netSavings = income - expenses - goalExpenses.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CashFlowProjectionPoint"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/plans/{planId}/pension/projection": {
      "parameters": [
        {
          "name": "planId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": [
          "Pension",
          "Analytics"
        ],
        "summary": "Pension projection with preserve-capital and spend-down variants",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PensionProjection"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CurrencyCode": {
        "type": "string",
        "minLength": 3,
        "maxLength": 3,
        "examples": [
          "RUB",
          "USD",
          "EUR"
        ]
      },
...
      "PensionSettings": {
        "type": "object",
        "required": [
          "currentAge",
          "retirementAge",
          "monthlyExpenses",
          "currency",
          "expectedReturnPct",
          "inflationPct",
          "withdrawalStrategy",
          "statePensionEnabled",
          "statePensionMonthly"
        ],
        "properties": {
          "currentAge": {
            "type": "integer",
            "minimum": 16,
            "maximum": 100,
            "examples": [
              32
            ]
          },
          "retirementAge": {
            "type": "integer",
            "minimum": 40,
            "maximum": 80,
            "examples": [
              60
            ]
          },
          "monthlyExpenses": {
            "type": "number",
            "minimum": 0,
            "examples": [
              120000
            ]
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyCode"
          },
          "expectedReturnPct": {
            "type": "number",
            "minimum": 0,
            "maximum": 30,
            "examples": [
              9
            ]
          },
          "inflationPct": {
            "type": "number",
            "minimum": 0,
            "maximum": 30,
            "examples": [
              7
            ]
          },
          "withdrawalStrategy": {
            "type": "string",
            "enum": [
              "preserve_capital",
              "spend_down_30y"
            ],
            "description": "UI scenarios: preserve / spend"
          },
          "statePensionEnabled": {
            "type": "boolean"
          },
          "statePensionMonthly": {
            "type": "number",
            "minimum": 0,
            "examples": [
              22000
            ]
          },
          "desiredMonthlyExpensesCurrentPrices": {
            "type": [
              "number",
              "null"
            ],
            "description": "Excel model row: desired retirement spending in current prices."
          },
          "projection": {
            "$ref": "#/components/schemas/PensionProjection"
          }
        }
      },
...
      "Contribution": {
        "type": "object",
        "required": [
          "id",
          "goalId",
          "amount",
          "date"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "goalId": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "minimum": 0
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyCode"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "note": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PlanState": {
        "type": "object",
        "required": [
          "id",
          "profile",
          "pension",
          "incomes",
          "expenses",
          "goals",
          "contributions",
          "budget"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/UserProfile"
          },
          "pension": {
            "$ref": "#/components/schemas/PensionSettings"
          },
          "incomes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IncomeSource"
            }
          },
          "expenses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpenseItem"
            }
          },
          "goals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Goal"
            }
          },
          "contributions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Contribution"
            }
          },
          "budget": {
            "$ref": "#/components/schemas/BudgetSettings"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "modelAssumptions": {
            "$ref": "#/components/schemas/ModelAssumptions"
          }
        }
      },
      "YearlyProjectionPoint": {
        "type": "object",
        "required": [
          "year",
          "income",
          "expenses",
          "goalsCost",
          "netSavings"
        ],
        "properties": {
          "year": {
            "type": "integer"
          },
          "income": {
            "type": "number"
          },
          "expenses": {
            "type": "number"
          },
          "goalsCost": {
            "type": "number"
          },
          "netSavings": {
            "type": "number"
          }
        }
      },
      "DashboardMetrics": {
        "type": "object",
        "required": [
          "totalMonthlyIncome",
          "totalYearlyIncome",
          "totalMonthlyExpenses",
          "totalYearlyExpenses",
          "netMonthlyBalance",
          "netYearlyBalance",
          "savingsRatePct",
          "totalGoalsCost",
          "totalGoalsSaved",
          "totalGoalsRemaining",
          "monthlyGoalContribution",
          "availableForPension",
          "projectedPensionCapital",
          "yearsToRetirement",
          "emergencyFundTarget",
          "emergencyFundCurrent",
          "emergencyFundPct",
          "yearlyProjection"
        ],
        "properties": {
          "totalMonthlyIncome": {
            "type": "number"
          },
          "totalYearlyIncome": {
            "type": "number"
          },
          "totalMonthlyExpenses": {
            "type": "number"
          },
          "totalYearlyExpenses": {
            "type": "number"
          },
          "netMonthlyBalance": {
            "type": "number"
          },
          "netYearlyBalance": {
            "type": "number"
          },
          "savingsRatePct": {
            "type": "number"
          },
          "totalGoalsCost": {
            "type": "number"
          },
          "totalGoalsSaved": {
            "type": "number"
          },
          "totalGoalsRemaining": {
            "type": "number"
          },
          "monthlyGoalContribution": {
            "type": "number"
          },
          "availableForPension": {
            "type": "number"
          },
          "projectedPensionCapital": {
            "type": "number"
          },
          "yearsToRetirement": {
            "type": "integer"
          },
          "emergencyFundTarget": {
            "type": "number"
          },
          "emergencyFundCurrent": {
            "type": "number"
          },
          "emergencyFundPct": {
            "type": "number"
          },
          "yearlyProjection": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/YearlyProjectionPoint"
            }
          }
        }
      },
...
      "Scenario": {
        "type": "object",
        "required": [
          "id",
          "name",
          "emoji",
          "description",
          "createdAt",
          "updatedAt",
          "isBase"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "emoji": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isBase": {
            "type": "boolean"
          },
          "basePlanId": {
            "type": "string"
          },
          "snapshot": {
            "$ref": "#/components/schemas/PlanState"
          },
          "adjustments": {
            "type": "object",
            "properties": {
              "incomeAdjPct": {
                "type": "number"
              },
              "expenseAdjPct": {
                "type": "number"
              },
              "returnAdjPct": {
                "type": "number"
              },
              "inflationAdjPct": {
                "type": "number"
              },
              "retirementAgeShift": {
                "type": "integer"
              },
              "goalsCostAdjPct": {
                "type": "number"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ModelAssumptions": {
        "type": "object",
        "required": [
          "startYear",
          "monthsPerYear",
          "currency",
          "investmentReturnPct",
          "inflationSchedule"
        ],
        "properties": {
          "startYear": {
            "type": "integer",
            "examples": [
              2024
            ]
          },
          "projectionEndYear": {
            "type": [
              "integer",
              "null"
            ],
            "examples": [
              2076
            ]
          },
          "horizonYears": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 1,
            "maximum": 80,
            "examples": [
              53
            ]
          },
          "birthYear": {
            "type": [
              "integer",
              "null"
            ],
            "examples": [
              1993
            ]
          },
          "monthsPerYear": {
            "type": "integer",
            "minimum": 1,
            "maximum": 12,
            "default": 12
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyCode"
          },
          "initialCapital": {
            "type": "number",
            "default": 0
          },
          "investmentReturnPct": {
            "type": "number",
            "description": "Nominal annual return in percent points. Excel sample: 6."
          },
          "inflationSchedule": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/YearRatePoint"
            }
          },
          "sourceModel": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional import/source marker, e.g. uploaded Excel model checksum."
          }
        }
      },
      "BalanceSnapshot": {
        "type": "object",
        "required": [
          "year",
          "monthlyIncome",
          "yearlyIncome",
          "totalIncome",
          "monthlyExpenses",
          "yearlyExpenses",
          "goalExpenses",
          "totalOutflow",
          "netSavings"
        ],
        "properties": {
          "year": {
            "type": "integer"
          },
          "monthlyIncome": {
            "type": "number"
          },
          "yearlyIncome": {
            "type": "number"
          },
          "totalIncome": {
            "type": "number"
          },
          "monthlyExpenses": {
            "type": "number"
          },
          "yearlyExpenses": {
            "type": "number"
          },
          "monthlyGoalExpenses": {
            "type": "number"
          },
          "yearlyGoalExpenses": {
            "type": "number"
          },
          "goalExpenses": {
            "type": "number"
          },
          "totalOutflow": {
            "type": "number"
          },
          "netSavings": {
            "type": "number"
          }
        }
      },
      "CashFlowProjectionPoint": {
        "type": "object",
        "required": [
          "year",
          "age",
          "periodNo",
          "totalIncome",
          "totalExpenses",
          "totalGoalExpenses",
          "netSavings",
          "capitalEndOfYear"
        ],
        "properties": {
          "year": {
            "type": "integer"
          },
          "age": {
            "type": [
              "integer",
              "null"
            ]
          },
          "periodNo": {
            "type": "integer"
          },
          "monthlyIncome": {
            "type": "number"
          },
          "yearlyIncome": {
            "type": "number"
          },
          "totalIncome": {
            "type": "number"
          },
          "monthlyExpenses": {
            "type": "number"
          },
          "yearlyExpenses": {
            "type": "number"
          },
          "totalExpenses": {
            "type": "number"
          },
          "monthlyGoalExpenses": {
            "type": "number"
          },
          "yearlyGoalExpenses": {
            "type": "number"
          },
          "totalGoalExpenses": {
            "type": "number"
          },
          "netSavings": {
            "type": "number"
          },
          "investmentReturnPct": {
            "type": "number"
          },
          "capitalStartOfYear": {
            "type": "number"
          },
          "capitalEndOfYear": {
            "type": "number"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Validation or domain error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid access token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  },
...
}